@@ -20,6 +20,7 @@
     "@docusaurus/plugin-ideal-image": "^3.1.1",
     "@docusaurus/preset-classic": "^3.1.1",
     "@docusaurus/remark-plugin-npm2yarn": "^3.1.1",
+    "@docusaurus/theme-classic": "^3.1.1",
     "@docusaurus/theme-common": "^3.1.1",
     "@docusaurus/utils": "^3.1.1",
     "@fortawesome/fontawesome-svg-core": "^6.5.1",
@@ -30,12 +31,12 @@
     "@iconify/react": "^4.1.1",
     "@mdx-js/react": "^3.0.1",
     "@popperjs/core": "^2.11.8",
-    "@swc/core": "^1.4.1",
-    "@types/react-dom": "^18.2.19",
+    "@swc/core": "^1.4.8",
+    "@types/react-dom": "^18.2.22",
     "clsx": "^2.1.0",
     "docusaurus-plugin-plausible": "^0.0.5",
     "docusaurus-plugin-sass": "^0.2.5",
-    "docusaurus-remark-plugin-tab-blocks": "^3.0.0",
+    "docusaurus-remark-plugin-tab-blocks": "^3.1.0",
     "prism-react-renderer": "^2.3.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
@@ -45,7 +46,7 @@
     "remark-definition-list": "^2.0.0",
     "remark-deflist": "^1.0.0",
     "remark-math": "6",
-    "sass": "^1.70.0",
+    "sass": "^1.72.0",
     "swc-loader": "^0.2.6"
   },
   "devDependencies": {
@@ -53,25 +54,25 @@
     "@docusaurus/tsconfig": "^3.1.1",
     "@docusaurus/types": "^3.1.1",
     "@types/jest": "^29.5.12",
-    "@types/node": "^20.11.17",
-    "@types/react": "^18.2.55",
-    "@typescript-eslint/eslint-plugin": "7.0.1",
-    "@typescript-eslint/parser": "7.0.1",
-    "eslint": "8.56.0",
+    "@types/node": "^20.11.30",
+    "@types/react": "^18.2.70",
+    "@typescript-eslint/eslint-plugin": "7.3.1",
+    "@typescript-eslint/parser": "7.3.1",
+    "eslint": "8.57.0",
     "eslint-config-prettier": "9.1.0",
     "eslint-plugin-formatjs": "4.12.2",
     "eslint-plugin-jsx-a11y": "6.8.0",
     "eslint-plugin-prettier": "5.1.3",
-    "eslint-plugin-react": "7.33.2",
+    "eslint-plugin-react": "7.34.1",
     "eslint-plugin-react-hooks": "4.6.0",
-    "eslint-plugin-regexp": "2.2.0",
+    "eslint-plugin-regexp": "2.4.0",
     "jest": "^29.7.0",
     "prettier": "3.2.5",
     "prettier-plugin-css-order": "2.0.1",
     "prop-types": "^15.8.1",
     "react-waypoint": "^10.3.0",
-    "typescript": "^5.3.3",
-    "webpack": "^5.90.1"
+    "typescript": "^5.4.3",
+    "webpack": "^5.91.0"
   },
   "browserslist": {
     "production": [
